refactor synthesis stuff into its own subpackage and update fn calls after `repackage` PR
