+++ synthesis/synthesis.go
@@ -1,12 +1,15 @@
-package poly
+package synthesis
 
 import (
 	"errors"
-	"github.com/jmoiron/sqlx"
-	_ "github.com/mattn/go-sqlite3"
 	"regexp"
 	"strings"
 	"sync"
+
+	"github.com/TimothyStiles/poly/transform"
+	"github.com/TimothyStiles/poly/transform/codon"
+	"github.com/jmoiron/sqlx"
+	_ "github.com/mattn/go-sqlite3"
 )
 
 /******************************************************************************
@@ -66,7 +69,7 @@ func RemoveSequence(sequencesToRemove []string) func(string, chan DnaSuggestion,
 	return func(sequence string, c chan DnaSuggestion, wg *sync.WaitGroup) {
 		var enzymes []string
 		for _, enzyme := range sequencesToRemove {
-			enzymes = []string{enzyme, ReverseComplement(enzyme)}
+			enzymes = []string{enzyme, transform.ReverseComplement(enzyme)}
 			for _, site := range enzymes {
 				re := regexp.MustCompile(site)
 				locs := re.FindAllStringIndex(sequence, -1)
@@ -130,7 +133,7 @@ func findProblems(sequence string, problematicSequenceFuncs []func(string, chan 
 }
 
 // FixCds fixes a CDS given the CDS sequence, a codon table, and a list of functions to solve for.
-func FixCds(sqlitePath string, sequence string, codontable CodonTable, problematicSequenceFuncs []func(string, chan DnaSuggestion, *sync.WaitGroup)) (string, error) {
+func FixCds(sqlitePath string, sequence string, codontable codon.Table, problematicSequenceFuncs []func(string, chan DnaSuggestion, *sync.WaitGroup)) (string, error) {
 	db := sqlx.MustConnect("sqlite3", sqlitePath)
 	createMemoryDbSql := `
 	CREATE TABLE codon (
@@ -277,7 +280,7 @@ func FixCds(sqlitePath string, sequence string, codontable CodonTable, problemat
 // finding homopolymers, finding repeats, and ensuring a normal range of GC
 // content. It also allows users to put in sequences that they do not wish to
 // occur within their CDS, like restriction enzyme cut sites.
-func FixCdsSimple(sequence string, codontable CodonTable, sequencesToRemove []string) (string, error) {
+func FixCdsSimple(sequence string, codontable codon.Table, sequencesToRemove []string) (string, error) {
 	var functions []func(string, chan DnaSuggestion, *sync.WaitGroup)
 	// Remove homopolymers
 	functions = append(functions, RemoveSequence([]string{"AAAAAAAA", "GGGGGGGG"}))
