+++ synthesis/synthesis_test.go
@@ -1,10 +1,14 @@
-package poly
+package synthesis
 
 import (
 	"fmt"
 	"strings"
 	"sync"
 	"testing"
+
+	"github.com/TimothyStiles/poly/io/genbank"
+	"github.com/TimothyStiles/poly/transform"
+	"github.com/TimothyStiles/poly/transform/codon"
 )
 
 /******************************************************************************
@@ -42,11 +46,14 @@ The following tests are built into DnaChisel:
 
 ******************************************************************************/
 
+var dataDir string = "../data/"
+
 func ExampleFixCds() {
 	bla := "ATGAGTATTCAACATTTCCGTGTCGCCCTTATTCCCTTTTTTGCGGCATTTTGCCTTCCTGTTTTTGCTCACCCAGAAACGCTGGTGAAAGTAAAAGATGCTGAAGATCAGTTGGGTGCACGAGTGGGTTACATCGAACTGGATCTCAACAGCGGTAAGATCCTTGAGAGTTTTCGCCCCGAAGAACGTTTTCCAATGATGAGCACTTTTAAAGTTCTGCTATGTGGCGCGGTATTATCCCGTATTGACGCCGGGCAAGAGCAACTCGGTCGCCGCATACACTATTCTCAGAATGACTTGGTTGAGTACTCACCAGTCACAGAAAAGCATCTTACGGATGGCATGACAGTAAGAGAATTATGCAGTGCTGCCATAACCATGAGTGATAACACTGCGGCCAACTTACTTCTGACAACGATCGGAGGACCGAAGGAGCTAACCGCTTTTTTGCACAACATGGGGGATCATGTAACTCGCCTTGATCGTTGGGAACCGGAGCTGAATGAAGCCATACCAAACGACGAGCGTGACACCACGATGCCTGTAGCAATGGCAACAACGTTGCGCAAACTATTAACTGGCGAACTACTTACTCTAGCTTCCCGGCAACAATTAATAGACTGGATGGAGGCGGATAAAGTTGCAGGACCACTTCTGCGCTCGGCCCTTCCGGCTGGCTGGTTTATTGCTGATAAATCTGGAGCCGGTGAGCGTGGGTCTCGCGGTATCATTGCAGCACTGGGGCCAGATGGTAAGCCCTCCCGTATCGTAGTTATCTACACGACGGGGAGTCAGGCAACTATGGATGAACGAAATAGACAGATCGCTGAGATAGGTGCCTCACTGATTAAGCATTGGTAA"
-	sequence := ReadGbk("data/ecoli-mg1655.gff")
-	codonTable := GetCodonTable(11)
-	optimizationTable := sequence.GetOptimizationTable(codonTable)
+	sequence := genbank.Read(dataDir + "ecoli-mg1655.gff")
+	codonTable := codon.GetCodonTable(11)
+	codingRegions := codon.GetCodingRegions(sequence)
+	optimizationTable := codonTable.OptimizeTable(codingRegions)
 
 	var functions []func(string, chan DnaSuggestion, *sync.WaitGroup)
 	functions = append(functions, RemoveSequence([]string{"GAAGAC", "GGTCTC", "GCGATG", "CGTCTC", "GCTCTTC", "CACCTGC"}))
@@ -60,9 +67,10 @@ func ExampleFixCds() {
 func ExampleFixCdsSimple() {
 	bla := "ATGAAAAAAAAAAGTATTCAACATTTCCGTGTCGCCCTTATTCCCTTTTTTGCGGCATTTTGCCTTCCTGTTTTTGCTCACCCAGAAACGCTGGTGAAAGTAAAAGATGCTGAAGATCAGTTGGGTGCACGAGTGGGTTACATCGAACTGGATCTCAACAGCGGTAAGATCCTTGAGAGTTTTCGCCCCGAAGAACGTTTTCCAATGATGAGCACTTTTAAAGTTCTGCTATGTGGCGCGGTATTATCCCGTATTGACGCCGGGCAAGAGCAACTCGGTCGCCGCATACACTATTCTCAGAATGACTTGGTTGAGTACTCACCAGTCACAGAAAAGCATCTTACGGATGGCATGACAGTAAGAGAATTATGCAGTGCTGCCATAACCATGAGTGATAACACTGCGGCCAACTTACTTCTGACAACGATCGGAGGACCGAAGGAGCTAACCGCTTTTTTGCACAACATGGGGGATCATGTAACTCGCCTTGATCGTTGGGAACCGGAGCTGAATGAAGCCATACCAAACGACGAGCGTGACACCACGATGCCTGTAGCAATGGCAACAACGTTGCGCAAACTATTAACTGGCGAACTACTTACTCTAGCTTCCCGGCAACAATTAATAGACTGGATGGAGGCGGATAAAGTTGCAGGACCACTTCTGCGCTCGGCCCTTCCGGCTGGCTGGTTTATTGCTGATAAATCTGGAGCCGGTGAGCGTGGGTCTCGCGGTATCATTGCAGCACTGGGGCCAGATGGTAAGCCCTCCCGTATCGTAGTTATCTACACGACGGGGAGTCAGGCAACTATGGATGAACGAAATAGACAGATCGCTGAGATAGGTGCCTCACTGATTAAGCATTGGTAA"
 
-	sequence := ReadGbk("data/ecoli-mg1655.gff")
-	codonTable := GetCodonTable(11)
-	optimizationTable := sequence.GetOptimizationTable(codonTable)
+	sequence := genbank.Read(dataDir + "ecoli-mg1655.gff")
+	codonTable := codon.GetCodonTable(11)
+	codingRegions := codon.GetCodingRegions(sequence)
+	optimizationTable := codonTable.OptimizeTable(codingRegions)
 
 	fixedSeq, _ := FixCdsSimple(bla, optimizationTable, []string{"GGTCTC"})
 	fmt.Println(fixedSeq)
@@ -71,11 +79,11 @@ func ExampleFixCdsSimple() {
 }
 
 func TestFixCds(t *testing.T) {
-	ct := ReadCodonJSON("data/pichiaTable.json")
+	ct := codon.ReadCodonJSON(dataDir + "pichiaTable.json")
 	phusion := "MGHHHHHHHHHHSSGILDVDYITEEGKPVIRLFKKENGKFKIEHDRTFRPYIYALLRDDSKIEEVKKITGERHGKIVRIVDVEKVEKKFLGKPITVWKLYLEHPQDVPTIREKVREHPAVVDIFEYDIPFAKRYLIDKGLIPMEGEEELKILAFDIETLYHEGEEFGKGPIIMISYADENEAKVITWKNIDLPYVEVVSSEREMIKRFLRIIREKDPDIIVTYNGDSFDFPYLAKRAEKLGIKLTIGRDGSEPKMQRIGDMTAVEVKGRIHFDLYHVITRTINLPTYTLEAVYEAIFGKPKEKVYADEIAKAWESGENLERVAKYSMEDAKATYELGKEFLPMEIQLSRLVGQPLWDVSRSSTGNLVEWFLLRKAYERNEVAPNKPSEEEYQRRLRESYTGGFVKEPEKGLWENIVYLDFRALYPSIIITHNVSPDTLNLEGCKNYDIAPQVGHKFCKDIPGFIPSLLGHLLEERQKIKTKMKETQDPIEKILLDYRQKAIKLLANSFYGYYGYAKARWYCKECAESVTAWGRKYIELVWKELEEKFGFKVLYIDTDGLYATIPGGESEEIKKKALEFVKYINSKLPGLLELEYEGFYKRGFFVTKKRYAVIDEEGKVITRGLEIVRRDWSEIAKETQARVLETILKHGDVEEAVRIVKEVIQKLANYEIPPEKLAIYEQITRPLHEYKAIGPHVAVAKKLAAKGVKIKPGMVIGYIVLRGDGPISNRAILAEEYDPKKHKYDAEYYIENQVLPAVLRILEGFGYRKEDLRYQKTRQVGLTSWLNIKKSGTGGGGATVKFKYKGEEKEVDISKIKKVWRVGKMISFTYDEGGGKTGRGAVSEKDAPKELLQMLEKQKK*"
 	var functions []func(string, chan DnaSuggestion, *sync.WaitGroup)
 	functions = append(functions, RemoveSequence([]string{"GAAGAC", "GGTCTC", "GCGATG", "CGTCTC", "GCTCTTC", "CACCTGC"}))
-	seq, _ := Optimize(phusion, ct)
+	seq, _ := codon.Optimize(phusion, ct)
 	optimizedSeq, err := FixCds(":memory:", seq, ct, functions)
 	if err != nil {
 		fmt.Println(err)
@@ -88,7 +96,7 @@ func TestFixCds(t *testing.T) {
 		if strings.Contains(optimizedSeq, cutSite) {
 			t.Errorf("phusion" + " contains " + cutSite)
 		}
-		if strings.Contains(ReverseComplement(optimizedSeq), cutSite) {
+		if strings.Contains(transform.ReverseComplement(optimizedSeq), cutSite) {
 			t.Errorf("phusion" + " reverse complement contains " + cutSite)
 		}
 	}
